fixed galla(e)ry typo in model & controller and unified responses
