+++ models/galleryimage.js
@@ -1,0 +1,26 @@
+'use strict';
+const {
+    Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+    class GalleryImage extends Model {
+        /**
+         * Helper method for defining associations.
+         * This method is not a part of Sequelize lifecycle.
+         * The `models/index` file will call this method automatically.
+         */
+        static associate(models) {
+            // define association here
+        }
+    }
+
+    GalleryImage.init({
+        link: DataTypes.TEXT,
+        title: DataTypes.STRING,
+        description: DataTypes.TEXT
+    }, {
+        sequelize,
+        modelName: 'GalleryImage',
+    });
+    return GalleryImage;
+};
