+++ routes/galleryRoutes.js
@@ -1,4 +1,4 @@
-const controller = require("../controllers/gallaryController");
+const controller = require("../controllers/galleryController");
 const express = require("express");
 const router = express.Router();
 
@@ -7,7 +7,7 @@ const {isAdmin} = require('../middleware/roleAuthorization/role')
 const {validateAddImage, validateRemoveImage} = require("../middleware/requestValidators/gallary/gallaryRequestsValidator");
 router.use(passport.authenticate('jwt', {session: false}))
 
-router.get('/',controller.getGallaryImages)
+router.get('/',controller.getGalleryImages)
 router.post('/add', isAdmin,validateAddImage, controller.addImage);
 router.delete('/delete', isAdmin,validateRemoveImage, controller.removeImage)
