+++ routes/routes.js
@@ -12,7 +12,7 @@ const messageRouter = require("./messageRouter")
 const teacherRouter = require("./teacherRouter.js")
 const announcementRouter = require("./announcementRouter.js")
 const gradesRoutes = require("./gradesRoutes")
-const gallaryRoutes = require("./gallaryRoutes")
+const galleryRoutes = require("./galleryRoutes")
 
 router.use("/meeting",meetingRouter)
 router.use("/auth",authRoutes)
@@ -25,5 +25,5 @@ router.use("/teacher",teacherRouter)
 router.use("/grades",gradesRoutes)
 router.use("/announcement",announcementRouter)
 router.use("/nots", NotificationRouter)
-router.use("/gallary", gallaryRoutes)
+router.use("/gallery", galleryRoutes)
 module.exports = router
